requirements.txt
yt-dlp
openai-whisper
Mastodon.py
requests
Flask
Flask-HTTPAuth
ruff
